requirements.txt
# Packages légers et faciles à installer
opencv-python==4.8.1.78
numpy==1.24.3
psycopg2-binary==2.9.9
pyserial==3.5
python-dotenv==1.0.0
Pillow==10.1.0